fix: rename and move CompareGtfInfo
